refactor: improving external faucet screen
