+++ mobile/modules/ui-components/index.ts
@@ -4,3 +4,4 @@ export * from './templates'
 export * from './molecules'
 export * from './atoms'
 export * from './utils/format'
+export * from './utils'
